% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acknowledge.R
\name{Acknowledge}
\alias{Acknowledge}
\title{Acknowledge contributions to this project and package.}
\usage{
Acknowledge(a)
}
\value{
Acknowledgements.
}
\description{
Acknowledge contributions to this project and package.
}
\details{
This function acknowledges contributions to the Fingerprinting project.
}
\examples{
Acknowledge()

}
