% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process the raw accelerometry data}
\usage{
process(folder, demo)
}
\arguments{
\item{folder}{character, the folder that contains the raw data in csv files for the 32 subjects.}

\item{demo}{data.frame, the dataset containing subject demographics.}
}
\value{
One organized data object that contains walking data from left wrists that can be used in further visualization or analyses.
}
\description{
Process the raw accelerometry data
}
\details{
This function processes the raw accelerometry data.
}
\examples{
process("raw_data")

}
