% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start.R
\name{Start}
\alias{Start}
\title{Start introduction of the fingerprinting project and dataset used}
\usage{
Start()
}
\value{
general information of the fingerprinting project.
}
\description{
Step by step tutorial for downloading the dataset
}
\details{
This function introduces the fingerprinting project and dataset.
}
\examples{
Start()

}
