% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_3D.R
\name{visualize_3D}
\alias{visualize_3D}
\title{Visualize the translated and 3D data}
\usage{
visualize_3D(subject_id, df_fit, time_lags = c(0.15, 0.3))
}
\arguments{
\item{subject_id}{numeric, id of the subject.}

\item{df_fit}{tbl_df, the processed dataset obtained from process.}

\item{time_lags}{vector, a vector of time lags for signal translation.}
}
\value{
one figure containing 4 panels.
}
\description{
Visualize the translated and 3D data
}
\details{
This function creates one figure that shows translated walking data and corresponding 3D images of a subject for a specified time interval.
}
\examples{
visualize_3D(21, df_fit, c(0.15, 0.3))

}
