% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_raw.R
\name{visualize_raw}
\alias{visualize_raw}
\title{Visualize the raw and calculated data}
\usage{
visualize_raw(subject_data, seconds = 20)
}
\arguments{
\item{subject_data}{data.frame, the raw walking data for a subject.}

\item{seconds}{numeric, the number of seconds to show.}
}
\value{
one figure containing 4 panels.
}
\description{
Visualize the raw and calculated data
}
\details{
This function creates one figure that shows raw walking data and vector magnitude of a subject for a specified time interval.
}
\examples{
visualize_raw(subject_data, 20)

}
